requirements.txt
snscrape>=0.7.0
instaloader>=4.10
pandas>=2.0.0
reportlab>=4.0.0
tabulate>=0.9.0
Pillow>=10.0.0
transformers>=4.40.0
torch>=2.0.0
numpy>=1.25.0